fandomproject/making/cartoongan_pytorch_main/requirements.txt
torch==1.8.0
torchvision==0.9.0
pillow==9.0.0
black==21.12b0
flake8==4.0.1